[#9] improve handling of filesystem/os delay
This adds and check and wait loop for the task wrapper to read
the pickle parameter file. We already had one of these for the
existence of the pickle file, now we also check for correct
parsing of the pickle file (in case it's incomplete), and
delay/retry to give the FS a chance to stabilize.